+++ src/sanitizer/config.ts
@@ -4,6 +4,8 @@
 
 import { SanitizerOptions } from "../types.js";
 
+export const DEFAULT_MAX_INPUT_LENGTH = 1_000_000;
+
 /**
  * Default sanitizer options with minimal safe allowlists
  */
@@ -11,36 +13,59 @@ export const DEFAULT_OPTIONS: Required<SanitizerOptions> = {
   // Common HTML elements that are safe by default
   allowedTags: [
     // Headings
-    "h1", "h2", "h3", "h4", "h5", "h6",
-    
+    "h1",
+    "h2",
+    "h3",
+    "h4",
+    "h5",
+    "h6",
+
     // Basic text formatting
-    "p", "div", "span", "b", "i", "strong", "em", 
-    
+    "p",
+    "div",
+    "span",
+    "b",
+    "i",
+    "strong",
+    "em",
+
     // Links and media
-    "a", "img",
-    
+    "a",
+    "img",
+
     // Lists
-    "ul", "ol", "li", 
-    
+    "ul",
+    "ol",
+    "li",
+
     // Tables
-    "table", "tr", "td", "th", 
-    
+    "table",
+    "tr",
+    "td",
+    "th",
+
     // Other common elements
-    "br", "hr", "code", "pre", "blockquote"
+    "br",
+    "hr",
+    "code",
+    "pre",
+    "blockquote",
   ],
-  
+
   // Only the most essential attributes
   allowedAttributes: {
     // Links
     a: ["href", "target", "rel"],
-    
-    // Images 
+
+    // Images
     img: ["src", "alt", "width", "height"],
-    
+
     // Global attributes
-    "*": ["id", "class"]
+    "*": ["id", "class"],
   },
-  
+
+  maxInputLength: DEFAULT_MAX_INPUT_LENGTH,
+
   // Always self-close void elements
 };